Relocated multiaxisarm package
